updated code

+++ src/Ch03/03_04/Begin/CodeDemo.cpp
@@ -8,7 +8,7 @@
 
 int main(){
     cow my_cow("Hildy",7,pet);
-    std::cout << my_cow.get_name() << " is a type-" << (int)my_cow.get_purpose() << " cow." << std::endl;
+    std::cout << my_cow.get_name() << " is a type-" << my_cow.get_purpose() << " cow." << std::endl;
     std::cout << my_cow.get_name() << " is " << my_cow.get_age() << " years old." << std::endl;
     return (0);
 }
